// src/store/modules/projectposting/mutation-types.ts
export enum MutationTypes {
  PROJECTPOSTING_STEP_LOADING = "PROJECTPOSTING_STEP_LOADING",
  PROJECTPOSTING_STEP_LOADED = "PROJECTPOSTING_STEP_LOADED",
  PROJECTPOSTING_LOADING = "PROJECTPOSTING_LOADING",
  PROJECTPOSTING_LOADED = "PROJECTPOSTING_LOADED",
  CLEAR_CURRENT_PROJECTPOSTING = "CLEAR_CURRENT_PROJECTPOSTING",
  ADD_EMPLOYEE_LOADING = "ADD_EMPLOYEE_LOADING",
  ADD_EMPLOYEE_LOADED = "ADD_EMPLOYEE_LOADED",
  EMPLOYEES_LOADING = "EMPLOYEES_LOADING",
  EMPLOYEES_LOADED = "EMPLOYEES_LOADED",
}